requirements.txt
stremalit
azure-cognitiveservices-speech
python-dotenv